renamed view to viewlist
