--- src/components/pages/manage/View.js
@@ -1,42 +1,0 @@
-import React from 'react';
-
-const View = props => (
-  <div>
-    {props.data.map(item => (
-      <figure key={item.id}>
-        <p>{item.member_id}</p>
-      </figure>
-    ))}
-  </div>
-);
-
-export default View;
-//*******************EXAMPLE CODE
-// const RenderExampleListPage = props => (
-//   <div>
-//     {props.data.map(item => (
-//       <figure key={item.id}>
-//         <img src={item.thumbnailUrl} alt={item.title} />
-//         <figcaption>
-//           <h3>{item.title}</h3>
-//         </figcaption>
-//       </figure>
-//     ))}
-//   </div>
-// );
-
-// export default RenderExampleListPage;
-
-// // Don't forget your prop types! It will save you a lot of debugging headache as you add more features.
-// RenderExampleListPage.propTypes = {
-//   data: PropTypes.arrayOf(
-//     // Here is an example of enforcing an object structure that we expect to receive in our props:
-//     PropTypes.shape({
-//       // Here we require an id of type number or string to prevent a "unique key prop" warning
-//       id: PropTypes.oneOfType([PropTypes.number, PropTypes.string]).isRequired,
-//       title: PropTypes.string,
-//       url: PropTypes.string,
-//       thumbnailUrl: PropTypes.string,
-//     })
-//   ).isRequired,
-// };
